fixed colunms type on items(user_id) and images(item_id) tables

+++ database/migrations/2019_11_14_114331_create_images_table.php
@@ -15,7 +15,7 @@ class CreateImagesTable extends Migration
     {
         Schema::create('images', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('item_id')->unsigned();
+            $table->bigInteger('item_id')->unsigned();
             $table->string('location');
             $table->timestamps();
